Store predictions + eval
